clean up benchmarks
